react native restructure
