+++ src/pages/integration-builder/builder/frameworks/react-native/replacementCode/get-constructor-code-rn.ts
