+++ src/pages/integration-builder/builder/frameworks/react-native/replacementCode/get-module-import-rn.ts
