+++ src/pages/integration-builder/builder/frameworks/react-native/replacementCode/get-resolved-redirect-url.ts
