+++ src/pages/integration-builder/builder/frameworks/react-native/replacementCode/index.ts
