+++ src/pages/integration-builder/builder/frameworks/react-native/stepContent/instantiateSDK.mdx
