+++ src/pages/integration-builder/builder/frameworks/react-native/stepContent/platform-setup.mdx
