+++ src/pages/integration-builder/builder/frameworks/react-native/stepContent/register-app.mdx
